fixed proper hashing
