+++ grovedb/src/operations/delete.rs
@@ -64,7 +64,7 @@ impl GroveDb {
             }
         );
         let ops_len = ops.len();
-        dbg!(&ops);
+        // dbg!(&ops);
         self.apply_batch(ops, None, transaction)
             .map_ok(|_| ops_len as u16)
     }
