+++ grovedb/src/tests/mod.rs
@@ -1,3 +1,7 @@
+pub mod common;
+mod query_tests;
+mod tree_hashes_tests;
+
 use std::{
     ops::{Deref, DerefMut},
     option::Option::None,
@@ -5,7 +9,7 @@ use std::{
 
 use ::visualize::{visualize_stdout, Drawer, Visualize};
 use merk::{
-    tree::{combine_hash, value_hash},
+    tree::{combine_hash, kv_digest_to_kv_hash, node_hash, value_hash, NULL_HASH},
     CryptoHash,
 };
 use rand::Rng;
@@ -14,7 +18,8 @@ use tempfile::TempDir;
 use super::*;
 use crate::{
     query_result_type::QueryResultType::QueryKeyElementPairResultType,
-    reference_path::ReferencePathType, visualize::visualize_merk_stdout,
+    reference_path::ReferencePathType, tests::common::compare_result_tuples,
+    visualize::visualize_merk_stdout,
 };
 
 pub const TEST_LEAF: &[u8] = b"test_leaf";
@@ -2851,1316 +2856,6 @@ fn test_aux_with_transaction() {
     );
 }
 
-fn populate_tree_for_non_unique_range_subquery(db: &TempGroveDb) {
-    // Insert a couple of subtrees first
-    for i in 1985u32..2000 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
-        db.insert([TEST_LEAF], &i_vec, Element::empty_tree(), None, None)
-            .unwrap()
-            .expect("successful subtree insert");
-        // Insert element 0
-        // Insert some elements into subtree
-        db.insert(
-            [TEST_LEAF, i_vec.as_slice()],
-            b"\0",
-            Element::empty_tree(),
-            None,
-            None,
-        )
-        .unwrap()
-        .expect("successful subtree insert");
-
-        for j in 100u32..150 {
-            let mut j_vec = i_vec.clone();
-            j_vec.append(&mut (j as u32).to_be_bytes().to_vec());
-            db.insert(
-                [TEST_LEAF, i_vec.as_slice(), b"\0"],
-                &j_vec.clone(),
-                Element::new_item(j_vec),
-                None,
-                None,
-            )
-            .unwrap()
-            .expect("successful value insert");
-        }
-    }
-}
-
-fn populate_tree_for_non_unique_double_range_subquery(db: &TempGroveDb) {
-    // Insert a couple of subtrees first
-    for i in 0u32..10 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
-        db.insert([TEST_LEAF], &i_vec, Element::empty_tree(), None, None)
-            .unwrap()
-            .expect("successful subtree insert");
-        // Insert element 0
-        // Insert some elements into subtree
-        db.insert(
-            [TEST_LEAF, i_vec.as_slice()],
-            b"a",
-            Element::empty_tree(),
-            None,
-            None,
-        )
-        .unwrap()
-        .expect("successful subtree insert");
-
-        for j in 25u32..50 {
-            let j_vec = (j as u32).to_be_bytes().to_vec();
-            db.insert(
-                [TEST_LEAF, i_vec.as_slice(), b"a"],
-                &j_vec,
-                Element::empty_tree(),
-                None,
-                None,
-            )
-            .unwrap()
-            .expect("successful value insert");
-
-            // Insert element 0
-            // Insert some elements into subtree
-            db.insert(
-                [TEST_LEAF, i_vec.as_slice(), b"a", j_vec.as_slice()],
-                b"\0",
-                Element::empty_tree(),
-                None,
-                None,
-            )
-            .unwrap()
-            .expect("successful subtree insert");
-
-            for k in 100u32..110 {
-                let k_vec = (k as u32).to_be_bytes().to_vec();
-                db.insert(
-                    [TEST_LEAF, i_vec.as_slice(), b"a", &j_vec, b"\0"],
-                    &k_vec.clone(),
-                    Element::new_item(k_vec),
-                    None,
-                    None,
-                )
-                .unwrap()
-                .expect("successful value insert");
-            }
-        }
-    }
-}
-
-fn populate_tree_by_reference_for_non_unique_range_subquery(db: &TempGroveDb) {
-    // This subtree will be holding values
-    db.insert([TEST_LEAF], b"\0", Element::empty_tree(), None, None)
-        .unwrap()
-        .expect("successful subtree insert");
-
-    // This subtree will be holding references
-    db.insert([TEST_LEAF], b"1", Element::empty_tree(), None, None)
-        .unwrap()
-        .expect("successful subtree insert");
-    // Insert a couple of subtrees first
-    for i in 1985u32..2000 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
-        db.insert([TEST_LEAF, b"1"], &i_vec, Element::empty_tree(), None, None)
-            .unwrap()
-            .expect("successful subtree insert");
-        // Insert element 0
-        // Insert some elements into subtree
-        db.insert(
-            [TEST_LEAF, b"1", i_vec.as_slice()],
-            b"\0",
-            Element::empty_tree(),
-            None,
-            None,
-        )
-        .unwrap()
-        .expect("successful subtree insert");
-
-        for j in 100u32..150 {
-            let random_key = rand::thread_rng().gen::<[u8; 32]>();
-            let mut j_vec = i_vec.clone();
-            j_vec.append(&mut (j as u32).to_be_bytes().to_vec());
-
-            // We should insert every item to the tree holding items
-            db.insert(
-                [TEST_LEAF, b"\0"],
-                &random_key,
-                Element::new_item(j_vec.clone()),
-                None,
-                None,
-            )
-            .unwrap()
-            .expect("successful value insert");
-
-            db.insert(
-                [TEST_LEAF, b"1", i_vec.clone().as_slice(), b"\0"],
-                &random_key,
-                Element::new_reference(ReferencePathType::AbsolutePathReference(vec![
-                    TEST_LEAF.to_vec(),
-                    b"\0".to_vec(),
-                    random_key.to_vec(),
-                ])),
-                None,
-                None,
-            )
-            .unwrap()
-            .expect("successful value insert");
-        }
-    }
-}
-
-fn populate_tree_for_unique_range_subquery(db: &TempGroveDb) {
-    // Insert a couple of subtrees first
-    for i in 1985u32..2000 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
-        db.insert([TEST_LEAF], &i_vec, Element::empty_tree(), None, None)
-            .unwrap()
-            .expect("successful subtree insert");
-
-        db.insert(
-            [TEST_LEAF, &i_vec.clone()],
-            b"\0",
-            Element::new_item(i_vec),
-            None,
-            None,
-        )
-        .unwrap()
-        .expect("successful value insert");
-    }
-}
-
-fn populate_tree_by_reference_for_unique_range_subquery(db: &TempGroveDb) {
-    // This subtree will be holding values
-    db.insert([TEST_LEAF], b"\0", Element::empty_tree(), None, None)
-        .unwrap()
-        .expect("successful subtree insert");
-
-    // This subtree will be holding references
-    db.insert([TEST_LEAF], b"1", Element::empty_tree(), None, None)
-        .unwrap()
-        .expect("successful subtree insert");
-
-    for i in 1985u32..2000 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
-        db.insert([TEST_LEAF, b"1"], &i_vec, Element::empty_tree(), None, None)
-            .unwrap()
-            .expect("successful subtree insert");
-
-        // We should insert every item to the tree holding items
-        db.insert(
-            [TEST_LEAF, b"\0"],
-            &i_vec,
-            Element::new_item(i_vec.clone()),
-            None,
-            None,
-        )
-        .unwrap()
-        .expect("successful value insert");
-
-        // We should insert a reference to the item
-        db.insert(
-            [TEST_LEAF, b"1", i_vec.clone().as_slice()],
-            b"\0",
-            Element::new_reference(ReferencePathType::AbsolutePathReference(vec![
-                TEST_LEAF.to_vec(),
-                b"\0".to_vec(),
-                i_vec.clone(),
-            ])),
-            None,
-            None,
-        )
-        .unwrap()
-        .expect("successful value insert");
-    }
-}
-
-fn populate_tree_for_unique_range_subquery_with_non_unique_null_values(db: &mut TempGroveDb) {
-    populate_tree_for_unique_range_subquery(db);
-    db.insert([TEST_LEAF], &[], Element::empty_tree(), None, None)
-        .unwrap()
-        .expect("successful subtree insert");
-    db.insert([TEST_LEAF, &[]], b"\0", Element::empty_tree(), None, None)
-        .unwrap()
-        .expect("successful subtree insert");
-    // Insert a couple of subtrees first
-    for i in 100u32..200 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
-        db.insert(
-            [TEST_LEAF, &[], b"\0"],
-            &i_vec,
-            Element::new_item(i_vec.clone()),
-            None,
-            None,
-        )
-        .unwrap()
-        .expect("successful value insert");
-    }
-}
-
-pub fn compare_result_tuples(
-    result_set: Vec<(Vec<u8>, Vec<u8>, CryptoHash)>,
-    expected_result_set: Vec<(Vec<u8>, Vec<u8>)>,
-) {
-    assert_eq!(expected_result_set.len(), result_set.len());
-    for i in 0..expected_result_set.len() {
-        assert_eq!(expected_result_set[i].0, result_set[i].0);
-        assert_eq!(expected_result_set[i].1, result_set[i].1);
-    }
-}
-
-fn compare_result_sets(elements: &Vec<Vec<u8>>, result_set: &Vec<(Vec<u8>, Vec<u8>, CryptoHash)>) {
-    for i in 0..elements.len() {
-        assert_eq!(
-            deserialize_and_extract_item_bytes(&result_set[i].1).unwrap(),
-            elements[i]
-        )
-    }
-}
-
-fn deserialize_and_extract_item_bytes(raw_bytes: &[u8]) -> Result<Vec<u8>, Error> {
-    let elem = Element::deserialize(raw_bytes)?;
-    return match elem {
-        Element::Item(item, _) => Ok(item),
-        _ => Err(Error::CorruptedPath("expected only item type")),
-    };
-}
-
-#[test]
-fn test_get_range_query_with_non_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range(1988_u32.to_be_bytes().to_vec()..1992_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 200);
-
-    let mut first_value = 1988_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1991_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 200);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_query_with_unique_subquery() {
-    let mut db = make_test_grovedb();
-    populate_tree_for_unique_range_subquery(&mut db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range(1988_u32.to_be_bytes().to_vec()..1992_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-
-    query.set_subquery_key(subquery_key);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 4);
-
-    let first_value = 1988_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 1991_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 4);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_query_with_unique_subquery_on_references() {
-    let db = make_test_grovedb();
-    populate_tree_by_reference_for_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec(), b"1".to_vec()];
-    let mut query = Query::new();
-    query.insert_range(1988_u32.to_be_bytes().to_vec()..1992_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-
-    query.set_subquery_key(subquery_key);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 4);
-
-    let first_value = 1988_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 1991_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 4);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_query_with_unique_subquery_with_non_unique_null_values() {
-    let mut db = make_test_grovedb();
-    populate_tree_for_unique_range_subquery_with_non_unique_null_values(&mut db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_all();
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-
-    query.set_subquery_key(subquery_key);
-
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.add_conditional_subquery(
-        QueryItem::Key(b"".to_vec()),
-        Some(b"\0".to_vec()),
-        Some(subquery),
-    );
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 115);
-
-    let first_value = 100_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 1999_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 115);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_query_with_unique_subquery_ignore_non_unique_null_values() {
-    let mut db = make_test_grovedb();
-    populate_tree_for_unique_range_subquery_with_non_unique_null_values(&mut db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_all();
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-
-    query.set_subquery_key(subquery_key);
-
-    let subquery = Query::new();
-
-    query.add_conditional_subquery(
-        QueryItem::Key(b"".to_vec()),
-        Some(b"\0".to_vec()),
-        Some(subquery),
-    );
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 15);
-
-    let first_value = 1985_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 1999_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 15);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_inclusive_query_with_non_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_inclusive(1988_u32.to_be_bytes().to_vec()..=1995_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 400);
-
-    let mut first_value = 1988_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1995_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 400);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_inclusive_query_with_non_unique_subquery_on_references() {
-    let db = make_test_grovedb();
-    populate_tree_by_reference_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec(), b"1".to_vec()];
-    let mut query = Query::new();
-    query.insert_range_inclusive(1988_u32.to_be_bytes().to_vec()..=1995_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 400);
-
-    let mut first_value = 1988_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    // using contains as the elements get stored at random key locations
-    // hence impossible to predict the final location
-    // but must exist
-    assert!(elements.contains(&first_value));
-
-    let mut last_value = 1995_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert!(elements.contains(&last_value));
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 400);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_inclusive_query_with_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_inclusive(1988_u32.to_be_bytes().to_vec()..=1995_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-
-    query.set_subquery_key(subquery_key);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 8);
-
-    let first_value = 1988_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 1995_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 8);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_from_query_with_non_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_from(1995_u32.to_be_bytes().to_vec()..);
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 250);
-
-    let mut first_value = 1995_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1999_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 250);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_from_query_with_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_from(1995_u32.to_be_bytes().to_vec()..);
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-
-    query.set_subquery_key(subquery_key);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 5);
-
-    let first_value = 1995_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 1999_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 5);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_to_query_with_non_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_to(..1995_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 500);
-
-    let mut first_value = 1985_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1994_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 500);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_to_query_with_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_to(..1995_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-
-    query.set_subquery_key(subquery_key);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 10);
-
-    let first_value = 1985_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 1994_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 10);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_to_inclusive_query_with_non_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_to_inclusive(..=1995_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 550);
-
-    let mut first_value = 1985_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1995_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 550);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_to_inclusive_query_with_non_unique_subquery_and_key_out_of_bounds() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new_with_direction(false);
-    query.insert_range_to_inclusive(..=5000_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new_with_direction(false);
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 750);
-
-    let mut first_value = 1999_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1985_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 750);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_to_inclusive_query_with_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_to_inclusive(..=1995_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-
-    query.set_subquery_key(subquery_key);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 11);
-
-    let first_value = 1985_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 1995_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 11);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_after_query_with_non_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_after(1995_u32.to_be_bytes().to_vec()..);
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 200);
-
-    let mut first_value = 1996_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1999_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 200);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_after_to_query_with_non_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_after_to(1995_u32.to_be_bytes().to_vec()..1997_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 50);
-
-    let mut first_value = 1996_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1996_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 50);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_after_to_inclusive_query_with_non_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_after_to_inclusive(
-        1995_u32.to_be_bytes().to_vec()..=1997_u32.to_be_bytes().to_vec(),
-    );
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 100);
-
-    let mut first_value = 1996_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1997_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 100);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_after_to_inclusive_query_with_non_unique_subquery_and_key_out_of_bounds() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new_with_direction(false);
-    query.insert_range_after_to_inclusive(
-        1995_u32.to_be_bytes().to_vec()..=5000_u32.to_be_bytes().to_vec(),
-    );
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new_with_direction(false);
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key);
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 200);
-
-    let mut first_value = 1999_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1996_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 200);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_inclusive_query_with_double_non_unique_subquery() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_double_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new();
-    query.insert_range_inclusive((3u32).to_be_bytes().to_vec()..=(4u32).to_be_bytes().to_vec());
-
-    query.set_subquery_key(b"a".to_vec());
-
-    let mut subquery = Query::new();
-    subquery
-        .insert_range_inclusive((29u32).to_be_bytes().to_vec()..=(31u32).to_be_bytes().to_vec());
-
-    subquery.set_subquery_key(b"\0".to_vec());
-
-    let mut subsubquery = Query::new();
-    subsubquery.insert_all();
-
-    subquery.set_subquery(subsubquery);
-
-    query.set_subquery(subquery);
-
-    let path_query = PathQuery::new_unsized(path, query.clone());
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 60);
-
-    let first_value = 100_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 109_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 60);
-    compare_result_sets(&elements, &result_set);
-}
-
-#[test]
-fn test_get_range_query_with_limit_and_offset() {
-    let db = make_test_grovedb();
-    populate_tree_for_non_unique_range_subquery(&db);
-
-    let path = vec![TEST_LEAF.to_vec()];
-    let mut query = Query::new_with_direction(true);
-    query.insert_range(1990_u32.to_be_bytes().to_vec()..1995_u32.to_be_bytes().to_vec());
-
-    let subquery_key: Vec<u8> = b"\0".to_vec();
-    let mut subquery = Query::new();
-    subquery.insert_all();
-
-    query.set_subquery_key(subquery_key.clone());
-    query.set_subquery(subquery.clone());
-
-    // Baseline query: no offset or limit + left to right
-    let path_query = PathQuery::new(path.clone(), SizedQuery::new(query.clone(), None, None));
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 250);
-
-    let mut first_value = 1990_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1994_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 250);
-    compare_result_sets(&elements, &result_set);
-
-    subquery.left_to_right = false;
-
-    query.set_subquery_key(subquery_key.clone());
-    query.set_subquery(subquery.clone());
-
-    query.left_to_right = false;
-
-    // Baseline query: no offset or limit + right to left
-    let path_query = PathQuery::new(path.clone(), SizedQuery::new(query.clone(), None, None));
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 250);
-
-    let mut first_value = 1994_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 149_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1990_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 250);
-    compare_result_sets(&elements, &result_set);
-
-    subquery.left_to_right = true;
-
-    query.set_subquery_key(subquery_key.clone());
-    query.set_subquery(subquery.clone());
-
-    query.left_to_right = true;
-
-    // Limit the result to just 55 elements
-    let path_query = PathQuery::new(path.clone(), SizedQuery::new(query.clone(), Some(55), None));
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 55);
-
-    let mut first_value = 1990_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 100_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    // Second tree 5 element [100, 101, 102, 103, 104]
-    let mut last_value = 1991_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 104_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 55);
-    compare_result_sets(&elements, &result_set);
-
-    query.set_subquery_key(subquery_key.clone());
-    query.set_subquery(subquery.clone());
-
-    // Limit the result set to 60 elements but skip the first 14 elements
-    let path_query = PathQuery::new(
-        path.clone(),
-        SizedQuery::new(query.clone(), Some(60), Some(14)),
-    );
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 60);
-
-    // Skips the first 14 elements, starts from the 15th
-    // i.e skips [100 - 113] starts from 114
-    let mut first_value = 1990_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 114_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    // Continues for 60 iterations
-    // Takes 36 elements from the first tree (50 - 14)
-    // takes the remaining 24 from the second three (60 - 36)
-    let mut last_value = 1991_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 123_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 60);
-    compare_result_sets(&elements, &result_set);
-
-    query.set_subquery_key(subquery_key.clone());
-    query.set_subquery(subquery.clone());
-
-    query.left_to_right = false;
-
-    // Limit the result set to 60 element but skip first 10 elements (this time
-    // right to left)
-    let path_query = PathQuery::new(
-        path.clone(),
-        SizedQuery::new(query.clone(), Some(60), Some(10)),
-    );
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 60);
-
-    // Skips the first 10 elements from the back
-    // last tree and starts from the 11th before the end
-    // Underlying subquery is ascending
-    let mut first_value = 1994_u32.to_be_bytes().to_vec();
-    first_value.append(&mut 110_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[0], first_value);
-
-    let mut last_value = 1993_u32.to_be_bytes().to_vec();
-    last_value.append(&mut 119_u32.to_be_bytes().to_vec());
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 60);
-    compare_result_sets(&elements, &result_set);
-
-    query.set_subquery_key(subquery_key.clone());
-    query.set_subquery(subquery.clone());
-
-    query.left_to_right = true;
-
-    // Offset bigger than elements in range
-    let path_query = PathQuery::new(
-        path.clone(),
-        SizedQuery::new(query.clone(), None, Some(5000)),
-    );
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 0);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 0);
-
-    query.set_subquery_key(subquery_key.clone());
-    query.set_subquery(subquery);
-
-    // Limit bigger than elements in range
-    let path_query = PathQuery::new(
-        path.clone(),
-        SizedQuery::new(query.clone(), Some(5000), None),
-    );
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 250);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 250);
-
-    // Test on unique subtree build
-    let db = make_test_grovedb();
-    populate_tree_for_unique_range_subquery(&db);
-
-    let mut query = Query::new_with_direction(true);
-    query.insert_range(1990_u32.to_be_bytes().to_vec()..2000_u32.to_be_bytes().to_vec());
-
-    query.set_subquery_key(subquery_key);
-
-    let path_query = PathQuery::new(path, SizedQuery::new(query.clone(), Some(5), Some(2)));
-
-    let (elements, _) = db
-        .query(&path_query, None)
-        .unwrap()
-        .expect("expected successful get_path_query");
-
-    assert_eq!(elements.len(), 5);
-
-    let first_value = 1992_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[0], first_value);
-
-    let last_value = 1996_u32.to_be_bytes().to_vec();
-    assert_eq!(elements[elements.len() - 1], last_value);
-
-    let proof = db.prove_query(&path_query).unwrap().unwrap();
-    let (hash, result_set) = GroveDb::verify_query(&proof, &path_query).unwrap();
-    assert_eq!(hash, db.root_hash(None).unwrap().unwrap());
-    assert_eq!(result_set.len(), 5);
-    compare_result_sets(&elements, &result_set);
-}
-
 #[test]
 fn test_root_hash() {
     let db = make_test_grovedb();
@@ -4372,71 +3067,4 @@ fn test_tree_value_exists_method_tx() {
         .expect("cannot commit transaction");
     assert!(db.has_raw([TEST_LEAF], b"key", None).unwrap().unwrap());
     assert!(db.has_raw([], b"leaf", None).unwrap().unwrap());
-}
-
-#[test]
-fn test_tree_value_hash_matches_underlying_merk_root_hash_empty_tree() {
-    let db = make_test_grovedb();
-
-    db.insert([TEST_LEAF], b"key1", Element::empty_tree(), None, None)
-        .unwrap()
-        .expect("successful subtree insert");
-
-    let test_leaf_merk = db
-        .open_non_transactional_merk_at_path([TEST_LEAF])
-        .unwrap()
-        .expect("should open merk");
-
-    let (elem_value, elem_value_hash) = test_leaf_merk
-        .get_value_and_value_hash(b"key1")
-        .unwrap()
-        .expect("should get value hash")
-        .expect("value hash should be some");
-
-    let underlying_merk = db
-        .open_non_transactional_merk_at_path([TEST_LEAF, b"key1"])
-        .unwrap()
-        .expect("should open merk");
-
-    let root_hash = underlying_merk.root_hash().unwrap();
-
-    let actual_value_hash = value_hash(&elem_value).unwrap();
-    let combined_value_hash = combine_hash(&actual_value_hash, &root_hash).unwrap();
-
-    assert_eq!(elem_value_hash, combined_value_hash);
-}
-
-#[test]
-fn test_tree_value_hash_matches_underlying_merk_root_hash() {
-    let db = make_test_grovedb();
-
-    db.insert([TEST_LEAF], b"key1", Element::empty_tree(), None, None)
-        .unwrap()
-        .expect("successful subtree insert");
-    db.insert([TEST_LEAF, b"key1"], b"key2", Element::new_item(vec![0]), None, None)
-        .unwrap()
-        .expect("successful subtree insert");
-
-    let test_leaf_merk = db
-        .open_non_transactional_merk_at_path([TEST_LEAF])
-        .unwrap()
-        .expect("should open merk");
-
-    let (elem_value, elem_value_hash) = test_leaf_merk
-        .get_value_and_value_hash(b"key1")
-        .unwrap()
-        .expect("should get value hash")
-        .expect("value hash should be some");
-
-    let underlying_merk = db
-        .open_non_transactional_merk_at_path([TEST_LEAF, b"key1"])
-        .unwrap()
-        .expect("should open merk");
-
-    let root_hash = underlying_merk.root_hash().unwrap();
-
-    let actual_value_hash = value_hash(&elem_value).unwrap();
-    let combined_value_hash = combine_hash(&actual_value_hash, &root_hash).unwrap();
-
-    assert_eq!(elem_value_hash, combined_value_hash);
 }
