+++ merk/src/tree/kv.rs
@@ -132,14 +132,18 @@ impl KV {
     /// Replaces the `KV`'s value with the given value and value hash,
     /// updates the hash and returns the modified `KV`.
     #[inline]
-    pub fn put_value_and_value_hash_then_update(
+    pub fn put_value_and_reference_value_hash_then_update(
         mut self,
         value: Vec<u8>,
-        value_hash: CryptoHash,
+        reference_value_hash: CryptoHash,
     ) -> CostContext<Self> {
         let mut cost = OperationCost::default();
+        let actual_value_hash = value_hash(value.as_slice()).unwrap_add_cost(&mut cost);
+        let combined_value_hash =
+            combine_hash(&actual_value_hash, &reference_value_hash).unwrap_add_cost(&mut cost);
+        //dbg!("combined_hash for propagation",std::str::from_utf8(value.as_slice()), hex::encode(actual_value_hash),hex::encode(combined_value_hash), hex::encode(reference_value_hash));
         self.value = value;
-        self.value_hash = value_hash;
+        self.value_hash = combined_value_hash;
         self.hash = kv_digest_to_kv_hash(self.key(), self.value_hash()).unwrap_add_cost(&mut cost);
         self.wrap_with_cost(cost)
     }
@@ -147,18 +151,14 @@ impl KV {
     /// Replaces the `KV`'s value with the given value and value hash,
     /// updates the hash and returns the modified `KV`.
     #[inline]
-    pub fn put_value_with_value_hash_and_value_cost_then_update(
+    pub fn put_value_with_reference_value_hash_and_value_cost_then_update(
         mut self,
         value: Vec<u8>,
-        value_hash: CryptoHash,
+        reference_value_hash: CryptoHash,
         value_cost: u32,
     ) -> CostContext<Self> {
-        let mut cost = OperationCost::default();
-        self.value = value;
-        self.value_hash = value_hash;
         self.value_defined_cost = Some(value_cost);
-        self.hash = kv_digest_to_kv_hash(self.key(), self.value_hash()).unwrap_add_cost(&mut cost);
-        self.wrap_with_cost(cost)
+        self.put_value_and_reference_value_hash_then_update(value,reference_value_hash)
     }
 
     /// Returns the key as a slice.
